refactor(website): share nav and footer between landing and docs
Pulls the nav and footer out of `home-landing.vue` into reusable
`bedrock-nav.vue` and `bedrock-footer.vue` components. Same chrome
(brand mark, theme toggle, GitHub CTA), per-page links and footer
columns passed in via props.

Also retargets the landing's `Docs` nav link and `Read the docs`
hero CTA at the new `/bedrock/introduction` entry point.
