+++ apps/website/.vitepress/theme/home-landing.vue
@@ -1,8 +1,10 @@
 <script setup lang="ts">
 import { onBeforeUnmount, ref } from "vue";
-import { useData } from "vitepress";
 
 import { version as bedrockVersion } from "@bedrock-rbx/core/package.json";
+
+import BedrockFooter, { type BedrockFooterColumn } from "./components/bedrock-footer.vue";
+import BedrockNav, { type BedrockNavLink } from "./components/bedrock-nav.vue";
 
 import configHtml from "../../landing/examples/config.ts?highlighted";
 import deployHtml from "../../landing/examples/deploy.ts?highlighted";
@@ -16,11 +18,62 @@ type TermId = "diff" | "deploy" | "migrate";
 const INSTALL_COMMAND = "pnpm add @bedrock-rbx/core";
 const COPIED_RESET_MS = 1200;
 
-const { isDark } = useData();
 const activeTab = ref<TabId>("config");
 const activeTerm = ref<TermId>("diff");
 const copyState = ref<"idle" | "copied">("idle");
 let copyResetTimer: ReturnType<typeof setTimeout> | undefined;
+
+const navLinks: ReadonlyArray<BedrockNavLink> = [
+	{ href: "#features", text: "Features" },
+	{ href: "#install", text: "Quickstart" },
+	{ href: "#ecosystem", text: "Ocale" },
+	{ href: "/bedrock/introduction", text: "Docs" },
+];
+
+const footerColumns: ReadonlyArray<BedrockFooterColumn> = [
+	{
+		links: [
+			{ href: "/bedrock/guide/getting-started", text: "Getting started" },
+			{ href: "#", text: "Configuration" },
+			{ href: "#", text: "CLI reference" },
+			{ href: "#", text: "Migration from Mantle" },
+		],
+		title: "Bedrock",
+	},
+	{
+		links: [
+			{ href: "/ocale/guide/getting-started", text: "Getting started" },
+			{ href: "/ocale/api/", text: "Resource clients" },
+			{ href: "/ocale/guide/errors", text: "Error hierarchy" },
+		],
+		title: "Ocale",
+	},
+	{
+		links: [
+			{
+				external: true,
+				href: "https://github.com/christopher-buss/bedrock",
+				text: "GitHub",
+			},
+			{
+				external: true,
+				href: "https://github.com/christopher-buss/bedrock/projects?query=is%3Aopen",
+				text: "Roadmap",
+			},
+			{
+				external: true,
+				href: "https://github.com/christopher-buss/bedrock/tree/main/docs/adr",
+				text: "ADRs",
+			},
+			{
+				external: true,
+				href: "https://github.com/christopher-buss/bedrock/releases",
+				text: "Changelog",
+			},
+		],
+		title: "Project",
+	},
+];
 
 async function copyInstall(): Promise<void> {
 	try {
@@ -98,58 +151,11 @@ function navigateTerminalTab(event: KeyboardEvent, fromId: TermId): void {
 	activeTerm.value = nextId;
 	focusTabButton("term", nextId);
 }
-
-function toggleTheme(): void {
-	isDark.value = !isDark.value;
-}
 </script>
 
 <template>
 	<div class="bedrock-landing">
-		<nav class="bedrock-nav">
-			<div class="wrap inner">
-				<a class="brand" href="#">
-					<span class="brand-mark"> <span /><span /><span /><span /> </span>
-					Bedrock<span class="nav-v">{{ VERSION }}</span>
-				</a>
-				<div class="nav-links">
-					<a href="#features">Features</a>
-					<a href="#install">Quickstart</a>
-					<a href="#ecosystem">Ocale</a>
-					<a href="/bedrock/guide/getting-started">Docs</a>
-				</div>
-				<button class="theme-toggle" aria-label="Toggle theme" @click="toggleTheme">
-					<svg v-if="!isDark" class="icon-moon" viewBox="0 0 16 16" fill="none">
-						<path
-							d="M13.5 9.5A5.5 5.5 0 0 1 6.5 2.5a5.5 5.5 0 1 0 7 7Z"
-							stroke="currentColor"
-							stroke-width="1.4"
-							stroke-linejoin="round"
-						/>
-					</svg>
-					<svg v-else class="icon-sun" viewBox="0 0 16 16" fill="none">
-						<circle cx="8" cy="8" r="3" stroke="currentColor" stroke-width="1.4" />
-						<path
-							d="M8 1.5v1.5M8 13v1.5M14.5 8H13M3 8H1.5M12.6 3.4l-1.1 1.1M4.5 11.5l-1.1 1.1M12.6 12.6l-1.1-1.1M4.5 4.5 3.4 3.4"
-							stroke="currentColor"
-							stroke-width="1.4"
-							stroke-linecap="round"
-						/>
-					</svg>
-				</button>
-				<a class="nav-cta" href="https://github.com/christopher-buss/bedrock">
-					GitHub
-					<svg width="11" height="11" viewBox="0 0 12 12" fill="none">
-						<path
-							d="M3 9L9 3M9 3H4M9 3V8"
-							stroke="currentColor"
-							stroke-width="1.5"
-							stroke-linecap="round"
-						/>
-					</svg>
-				</a>
-			</div>
-		</nav>
+		<BedrockNav :links="navLinks" />
 
 		<div class="wrap hero-wrap">
 			<section class="hero">
@@ -185,7 +191,7 @@ function toggleTheme(): void {
 									/>
 								</svg>
 							</a>
-							<a class="btn btn-ghost-dark" href="/bedrock/guide/getting-started">
+							<a class="btn btn-ghost-dark" href="/bedrock/introduction">
 								Read the docs
 							</a>
 						</div>
@@ -668,65 +674,7 @@ function toggleTheme(): void {
 			</div>
 		</section>
 
-		<footer class="bedrock-foot">
-			<div class="wrap">
-				<div class="foot">
-					<div class="foot-brand">
-						<a class="brand" href="/">
-							<span class="brand-mark"> <span /><span /><span /><span /> </span>
-							Bedrock
-						</a>
-						<p>Infrastructure-as-Code for Roblox. Typed, Open Cloud only.</p>
-					</div>
-					<div>
-						<h5>Bedrock</h5>
-						<ul>
-							<li><a href="/bedrock/guide/getting-started">Getting started</a></li>
-							<li><a href="#">Configuration</a></li>
-							<li><a href="#">CLI reference</a></li>
-							<li><a href="#">Migration from Mantle</a></li>
-						</ul>
-					</div>
-					<div>
-						<h5>Ocale</h5>
-						<ul>
-							<li><a href="/ocale/guide/getting-started">Getting started</a></li>
-							<li><a href="/ocale/api/">Resource clients</a></li>
-							<li><a href="/ocale/guide/errors">Error hierarchy</a></li>
-						</ul>
-					</div>
-					<div>
-						<h5>Project</h5>
-						<ul>
-							<li>
-								<a href="https://github.com/christopher-buss/bedrock">GitHub</a>
-							</li>
-							<li>
-								<a
-									href="https://github.com/christopher-buss/bedrock/projects?query=is%3Aopen"
-									>Roadmap</a
-								>
-							</li>
-							<li>
-								<a
-									href="https://github.com/christopher-buss/bedrock/tree/main/docs/adr"
-									>ADRs</a
-								>
-							</li>
-							<li>
-								<a href="https://github.com/christopher-buss/bedrock/releases"
-									>Changelog</a
-								>
-							</li>
-						</ul>
-					</div>
-				</div>
-				<div class="foot-bottom">
-					<div>&copy; 2026 &middot; MIT Licensed &middot; {{ VERSION }}</div>
-					<div>built with vitepress</div>
-				</div>
-			</div>
-		</footer>
+		<BedrockFooter :columns="footerColumns" />
 	</div>
 </template>
 
@@ -804,7 +752,6 @@ html.dark .bedrock-landing {
 }
 
 .bedrock-landing,
-.bedrock-landing .bedrock-nav,
 .bedrock-landing .hero {
 	transition:
 		background-color 0.25s var(--ease),
@@ -839,127 +786,6 @@ html.dark .bedrock-landing {
 	max-width: 1200px;
 	margin: 0 auto;
 	padding: 0 32px;
-}
-
-.bedrock-nav {
-	position: sticky;
-	top: 0;
-	z-index: 40;
-	backdrop-filter: blur(14px);
-	background: var(--bg-nav);
-	border-bottom: 1px solid var(--line);
-}
-
-.bedrock-nav .inner {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	height: 60px;
-}
-
-.nav-links {
-	display: flex;
-	gap: 28px;
-}
-
-.nav-links a {
-	font-size: 14px;
-	color: var(--ink-2);
-	transition: color 0.15s var(--ease);
-}
-
-.nav-links a:hover {
-	color: var(--ink);
-}
-
-.nav-cta {
-	font-size: 13px;
-	padding: 6px 12px 6px 14px;
-	border: 1px solid var(--line-strong);
-	border-radius: 999px;
-	display: inline-flex;
-	align-items: center;
-	gap: 6px;
-	transition: all 0.15s var(--ease);
-}
-
-.nav-cta:hover {
-	border-color: var(--ink);
-	background: var(--ink);
-	color: var(--bg);
-}
-
-.nav-v {
-	font-family: var(--f-mono);
-	font-size: 11px;
-	color: var(--ink-4);
-	padding: 2px 6px;
-	background: var(--bg-soft);
-	border-radius: 3px;
-	margin-left: 8px;
-}
-
-.theme-toggle {
-	width: 32px;
-	height: 32px;
-	border: 1px solid var(--line);
-	border-radius: 999px;
-	display: inline-flex;
-	align-items: center;
-	justify-content: center;
-	color: var(--ink-2);
-	transition: all 0.15s var(--ease);
-	margin-right: 6px;
-}
-
-.theme-toggle:hover {
-	border-color: var(--ink-3);
-	color: var(--ink);
-	background: var(--bg-soft);
-}
-
-.theme-toggle svg {
-	width: 14px;
-	height: 14px;
-}
-
-.brand {
-	display: inline-flex;
-	align-items: center;
-	gap: 10px;
-	font-family: var(--f-serif);
-	font-size: 23px;
-	letter-spacing: -0.01em;
-}
-
-.brand-mark {
-	width: 22px;
-	height: 22px;
-	display: grid;
-	grid-template-rows: repeat(4, 1fr);
-	gap: 2px;
-}
-
-.brand-mark span {
-	background: var(--ink);
-	border-radius: 1px;
-}
-
-.brand-mark span:nth-child(1) {
-	opacity: 0.3;
-}
-
-.brand-mark span:nth-child(2) {
-	opacity: 0.55;
-}
-
-.brand-mark span:nth-child(3) {
-	opacity: 0.8;
-}
-
-.brand-mark span:nth-child(4) {
-	background: var(--accent);
-	opacity: 1;
 }
 
 /* Eyebrow */
@@ -2126,77 +1952,13 @@ html.dark .bedrock-landing {
 	cursor: default;
 }
 
-/* Footer */
-.bedrock-foot {
-	padding: 72px 0 48px;
-	background: var(--bg-soft);
-	border-top: 1px solid var(--line);
-}
-
-.foot {
-	display: grid;
-	grid-template-columns: 1.4fr 1fr 1fr 1fr;
-	gap: 48px;
-	margin-bottom: 56px;
-}
-
-.foot h5 {
-	font-family: var(--f-mono);
-	font-size: 11px;
-	font-weight: 500;
-	letter-spacing: 0.14em;
-	text-transform: uppercase;
-	color: var(--ink-3);
-	margin: 0 0 16px;
-}
-
-.foot ul {
-	list-style: none;
-	padding: 0;
-	margin: 0;
-}
-
-.foot li {
-	margin-bottom: 10px;
-}
-
-.foot a {
-	font-size: 14px;
-	color: var(--ink-2);
-	transition: color 0.15s var(--ease);
-}
-
-.foot a:hover {
-	color: var(--ink);
-}
-
-.foot-brand p {
-	font-size: 14px;
-	color: var(--ink-3);
-	max-width: 34ch;
-	margin: 14px 0 0;
-	line-height: 1.55;
-}
-
-.foot-bottom {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	padding-top: 28px;
-	border-top: 1px solid var(--line);
-	font-size: 12px;
-	color: var(--ink-4);
-	font-family: var(--f-mono);
-}
-
 /* Responsive */
 @media (max-width: 960px) {
 	.hero-inner,
 	.tracks-head,
 	.tracks-grid,
 	.eco-head,
-	.install-grid,
-	.foot {
+	.install-grid {
 		grid-template-columns: 1fr;
 		gap: 40px;
 	}
@@ -2208,10 +1970,6 @@ html.dark .bedrock-landing {
 
 	.feature-grid {
 		grid-template-columns: 1fr 1fr;
-	}
-
-	.nav-links {
-		display: none;
 	}
 
 	.eco-layer {
